flask-api/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyJWT
requests
Pillow
